refactor(demo,website): extract shared sample-workflow fixtures to _utils
The Tabs addon demo reused SampleWorkflowModel/SampleWorkflowService
and their step components directly from the Wizard addon demo's
sample/ folder, the only cross-addon-demo reach-into in the app. Move
the shared model/service/steps into a new _utils/sample-workflow
folder (in both the demo app and website docs) so Wizard, Stepper,
and Tabs demos import a common location instead of one demo reaching
into another's internals.
