+++ projects/demo/src/app/addons/_utils/sample-workflow/steps/select-datastore/select-datastore.model.ts
