+++ projects/demo/src/app/addons/_utils/sample-workflow/steps/select-vm/select-vm.component.html
