+++ projects/website/src/app/documentation/demos/_utils/sample-workflow/sample-workflow.model.ts
