+++ projects/website/src/app/documentation/demos/_utils/sample-workflow/steps/object-name/object-name.model.ts
