+++ projects/website/src/app/documentation/demos/_utils/sample-workflow/steps/ready-to-complete/ready-to-complete.component.ts
