+++ projects/website/src/app/documentation/demos/_utils/sample-workflow/steps/select-datastore/select-datastore.component.html
