+++ projects/website/src/app/documentation/demos/_utils/sample-workflow/steps/select-vm/select-vm.component.html
